backend/requirements.txt
fastapi
uvicorn
langchain
openai
psycopg2-binary
python-dotenv
yfinance
supabase